removes global events and localizes them to modules
