--- src/PublicWebsite/src/messages/events.ts
@@ -1,11 +1,0 @@
-﻿
-export const Events = {
-  RoomTypeIdsAvailable: 'RoomTypeIdsAvailable',
-  RatesFetched: 'RatesFetched',
-  RatesSummaryFetched: 'RatesSummaryFetched',
-  BookRoom: 'BookRoom',
-  GuestSubmitted: 'GuestSubmitted',
-  ReservationSubmitted: 'ReservationSubmitted',
-  PaymentSubmitted: 'PaymentSubmitted'
-}
-
