+++ src/PublicWebsite/src/modules/rates/rates-summary.ts
@@ -1,8 +1,12 @@
 ﻿import {EventAggregator} from 'aurelia-event-aggregator'
 import {autoinject, bindable} from 'aurelia-framework';
 import {HttpClient} from 'aurelia-fetch-client';
-import {Events} from '../../messages/events';
 import shoppingCart from "../../shoppingCart";
+
+const Events = {
+  RatesFetched: 'RatesFetched',
+  RatesSummaryFetched: 'RatesSummaryFetched'
+}
 
 @autoinject()
 export class RatesSummary {
